requirements.txt
# Веб-фреймворк Flask
Flask==2.0.3

# Расширение для работы с формами WTForms
Flask-WTF==1.0.1

# Bootstrap для стилизации интерфейса
Flask-Bootstrap==3.3.7.1

# Утилиты для работы с файлами и путями
Werkzeug==2.0.3

# Библиотека для работы с изображениями
Pillow==9.0.1

# Библиотека для построения графиков
matplotlib==3.5.2

# Библиотека для численных вычислений
numpy==1.22.3

# WSGI-сервер для развёртывания приложения
gunicorn==20.1.0